Extract song list UI to SongListPane
Move the left-side search/filters/table/favorites UI out of SongsTab into a new SongListPane composable (composeApp/.../SongListPane.kt). SongsTab now delegates to SongListPane and forwards state and callbacks (search, sort, column resize/order, favorites, context menu, add-to-schedule, presenter actions, settings saves). This refactor improves modularity and readability without changing behaviour; settings and play-count wiring are preserved.
